backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.3.0
python-dotenv==1.0.1
openai-whisper==20231117
prometheus-client==0.20.0
gunicorn==22.0.0
requests==2.32.3  # For downloading the index
zipfile  # This is a standard library, but good to note

# LLM & RAG
google-generativeai==0.5.4
langchain==0.2.16
faiss-cpu==1.8.0
langchain-google-genai==1.0.6
langchain-community==0.2.12
langchain-ollama==0.1.3
nltk==3.9.1
qdrant-client==1.10.1
redis==5.0.1

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Migrations
alembic==1.13.2

# Dev Tools
ruff==0.6.9
pytest==8.3.3
httpx==0.27.2